% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{bird}
\alias{bird}
\title{Plot colour palette as a bird}
\usage{
bird(chrom)
}
\arguments{
\item{chrom}{A vector of colours}
}
\value{
A bird plot
}
\description{
The color for the bird were chosen from the ltc color palette
}
\examples{
paloma <- ltc("paloma")
bird(paloma)
}
