% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2transparent.R
\name{col2transparent}
\alias{col2transparent}
\title{col2transparent}
\usage{
col2transparent(col, alpha.perc = 0)
}
\arguments{
\item{col}{R color name or R color pal. This parameter is mandatory.}

\item{alpha.perc}{An alpha-transparency level in percent. Default value is 0.}
}
\value{
A character vector with elements of 9 characters, followed by the red, blue, green and alpha values in hexadecimal (after rescaling to 0 ... 255).
}
\description{
R color to transparent conversion.
}
\examples{
paloma <- ltc("paloma")
paloma <- col2transparent(paloma,50)
plts(paloma)
}
\seealso{
\code{\link[grDevices:palette]{grDevices::palette()}}
}
\author{
Kristian K Ullrich
}
