% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\docType{data}
\name{info}
\alias{info}
\title{Add information about the bio}
\format{
An object of class \code{data.frame} with 13 rows and 2 columns.
}
\usage{
info
}
\description{
Add information about the bio
}
\keyword{datasets}
