% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{ltc}
\alias{ltc}
\title{Select ltc palette}
\usage{
ltc(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{n}{Number of colors desired. Unfortunately most palettes now only
have 4 or 5 colors. But hopefully we'll add more palettes soon. All color
schemes are derived from the most excellent Tumblr blog:
If omitted, uses all colours.#' @return A vector of hex colour codes}
}
\description{
This function returns colour palettes
}
\examples{
ltc("paloma")
ltc("dora")
}
