% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{ltc_palette}
\alias{ltc_palette}
\title{Select ltc palette}
\usage{
ltc_palette(palette_name)
}
\arguments{
\item{palette_name}{Name of desired palette}
}
\value{
A vector of hex colour codes
}
\description{
This function returns colour palettes
}
\examples{
ltc_palette("paloma")
ltc_palette("dora")
}
