% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{List of colour palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
palettes
}
\description{
List of colour palettes
}
\keyword{datasets}
