% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{palt}
\alias{palt}
\title{Plot colour palette}
\usage{
palt(x)
}
\arguments{
\item{x}{A vector of colours}
}
\value{
An image showing colours in palette
}
\description{
Plot chosen colour palette from ltc package
}
\examples{
paloma <- ltc("paloma")
palt(paloma)
}
