% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{plts}
\alias{plts}
\title{Plot colour palette as sinus curve}
\usage{
plts(chromata, ...)
}
\arguments{
\item{chromata}{A vector of colours}
}
\value{
A sinus curve showing colours in palette
}
\description{
Plot chosen colour palette from ltc package as sinus curve
}
\examples{
paloma <- ltc("paloma")
plts(paloma)
}
