% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{print_pal}
\alias{print_pal}
\title{Plot colour palette}
\usage{
print_pal(x)
}
\arguments{
\item{x}{A vector of colours}
}
\value{
An image showing colours in palette
}
\description{
Plot chosen colour palette from ltc package
}
\examples{
paloma <- ltc_palette("paloma")
print_pal(paloma)
}
