% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{bird}
\alias{bird}
\title{Plot a Colour Palette as a Bird}
\usage{
bird(chrom)
}
\arguments{
\item{chrom}{A vector of colours from one of the \code{ltc} palettes.}
}
\value{
A ggplot2 object showing a bird drawing using the selected colours.
}
\description{
Visualizes a selected colour palette in the form of a bird drawing.
}
\examples{
\dontrun{
paloma <- ltc("paloma")
bird(paloma)
}
}
