% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\docType{data}
\name{info}
\alias{info}
\title{Information about the Colour Palettes}
\format{
An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
info
}
\description{
This dataframe contains the backstory or inspiration behind each color palette.
}
\keyword{datasets}
