% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\docType{data}
\name{ltc}
\alias{ltc}
\alias{palettes}
\title{List of colour palettes}
\format{
An object of class \code{list} of length 24.
}
\usage{
palettes

ltc(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Character. The name of the desired palette.}

\item{n}{Integer. The number of colors you want from the palette.
If omitted, it uses all colors from the palette.}

\item{type}{The type of palette. Either "discrete" or "continuous".}
}
\value{
A vector of hex color codes
}
\description{
A list containing predefined colour palettes with artistic backstories.

This function provides the desired colour palette by name.
}
\details{
ltc: A Collection of Art-inspired Colour Palettes

This package provides a collection of color palettes inspired by
art, nature, and personal preferences. Each palette has a backstory,
providing context and meaning to the colors.
}
\examples{
\dontrun{
ltc("paloma")
ltc("dora", n = 3)
}
}
\keyword{datasets}
