% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{pltc}
\alias{pltc}
\title{Plot a Colour Palette}
\usage{
pltc(chromata, ...)
}
\arguments{
\item{chromata}{A vector of colours from one of the \code{ltc} palettes.}

\item{...}{Additional arguments passed on to ggplot functions.}
}
\value{
A ggplot2 object showing the selected colours.
}
\description{
Visualizes a selected colour palette as a bar of colours.
}
\examples{
\dontrun{
paloma <- ltc("paloma")
pltc(paloma)
}
}
