% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltc_functions.R
\name{plts}
\alias{plts}
\title{Plot a Colour Palette as a Sinus Curve}
\usage{
plts(chromata, ...)
}
\arguments{
\item{chromata}{A vector of colours from one of the \code{ltc} palettes.}

\item{...}{Additional arguments passed on to ggplot functions.}
}
\value{
A sinusoidal curve with the selected colors.
}
\description{
Visualizes a selected colour palette as a sinusoidal curve.
}
\examples{
\dontrun{
paloma <- ltc("paloma")
plts(paloma)
}
}
